requirements.txt
# VisiHealth AI - Requirements
# Python >= 3.8

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Medical Image Processing
pydicom>=2.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
joblib>=1.3.0

# Knowledge Graph
networkx>=3.1

# Optional: Dataset tools
# fiftyone>=0.21.0

# Training utilities
tensorboard>=2.13.0
wandb>=0.15.0

# Evaluation
torchmetrics>=1.0.0
